% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_panel.R
\name{panel_cor}
\alias{panel_cor}
\alias{panel_idenDist}
\alias{panel_scatter_abline}
\alias{panel_scatter_abline_smallDots}
\alias{panel_scatter_abhvline}
\alias{panel_scatter_smallDots_abhvline}
\title{A better pairs plot function}
\usage{
panel_cor(x, y)

panel_idenDist(x, y)

panel_scatter_abline(x, y, a = 0, b = 1, col = "red")

panel_scatter_abline_smallDots(x, y, a = 0, b = 1, col = "red")

panel_scatter_abhvline(x, y)

panel_scatter_smallDots_abhvline(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{a}{a for abline}

\item{b}{b for abline}

\item{col}{col for abline}
}
\description{
scatter plot panel with the identity line

panel_scatter_abline with small dots

panel_scatter_abline with h and v lines at zero

panel_scatter_smallDots with h and v lines at zero
}
\examples{
pairs(iris[,-5], lower.panel = panel_cor)
pairs(iris[,-5],
upper.panel = panel_scatter_abline,
lower.panel = panel_idenDist)
pairs(iris[,-5],
lower.panel = panel_cor,
upper.panel = panel_scatter_abline)
pairs(iris[,-5],
lower.panel = panel_cor,
upper.panel = panel_scatter_abline_smallDots)
pairs(iris[,-5],
lower.panel = panel_cor,
upper.panel = panel_scatter_abhvline)
pairs(iris[,-5],
lower.panel = panel_cor,
upper.panel = panel_scatter_smallDots_abhvline)
}
\author{
Kevin Wang

Kevin Wang

Kevin Wang

Kevin Wang

Kevin Wang

Kevin Wang
}
