% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weights.R
\name{compute_weights}
\alias{compute_weights}
\title{Compute feature-wise concordance weights}
\usage{
compute_weights(x1, x2)
}
\arguments{
\item{x1}{A data matrix}

\item{x2}{A data matrix}
}
\value{
A vector
}
\description{
Compute feature-wise concordance weights
}
\examples{
n = 10
p = 10
x1 = matrix(rnorm(n * n, mean = 0, sd = 1), nrow = n, ncol = p)
x2 = matrix(rnorm(n * p, mean = 1, sd = 1), nrow = n, ncol = p)
compute_weights(x1, x2)
}
