% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\name{cost}
\alias{cost}
\alias{cost,cpop.class-method}
\title{A function for calculating the cost of a model fitted by cpop}
\usage{
cost(object)
}
\arguments{
\item{object}{An instance of an S4 class produced by \code{\link{cpop}}.}
}
\description{
Calculates the cost of a model fitted by cpop using the residual sum of squares and the penalty values.
}
