% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\name{CPOP.uneven}
\alias{CPOP.uneven}
\title{CPOP.uneven}
\usage{
CPOP.uneven(y, x, beta, sigsquared = 1)
}
\arguments{
\item{y}{A vector of length n containing the data.}

\item{x}{A vector of length n containing the locations of the data.}

\item{beta}{A positive real value for the penalty incurred for adding a changepoint (prevents over-fitting).}

\item{sigsquared}{Estimate of residual variance.}
}
\value{
List of minimum values and associated changepoints.
}
\description{
Algorithm for finding the best segmentation of data for a change-in-slope model where best is defined in terms of minimising
\eqn{ \frac{1}{\sigma^2} \sum_{i=1}^n (y_{i}-f_{i})^2+m\beta} 
where \eqn{y_i} are data, \eqn{f_i} is the fitted mean at point \eqn{i}, \eqn{m}  the number of changepoints
and \eqn{\beta} a parameter. (NOTE - this section  also forms part of the output produced by help in R - but the LaTeX does not get processed
unless the option \code{help_type="pdf"} is used to generate a pdf file - and this will only be the case if LaTeX is installed on the users system.
It is probably better to restrict the use of LaTeX to the cpop-package section of the documentation.
}
\examples{
ls()
 
}
\references{
Put any references here
}
