% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop1.R
\name{cpop1}
\alias{cpop1}
\alias{cpop1_iterate}
\title{Step 1 of the CPOP method}
\usage{
cpop1(
  z1,
  z2,
  y1,
  y2,
  w,
  family,
  n_iter = 20,
  alpha = 1,
  n_features = 50,
  s = "lambda.min",
  cpop1_method = "normal",
  ...
)

cpop1_iterate(
  z1,
  z2,
  y1,
  y2,
  w = NULL,
  family,
  s = "lambda.min",
  n_iter = 20,
  alpha = 1,
  n_features = 50,
  ...
)
}
\arguments{
\item{z1}{A data matrix, columns are pairwise-differences between the original data columns.}

\item{z2}{A data matrix, columns are pairwise-differences between the original data columns.
Column names should be identical to z1.}

\item{y1}{A vector of response variable. Same length as the number of rows of z1.}

\item{y2}{A vector of response variable. Same length as the number of rows of z2.}

\item{w}{A vector of weights.}

\item{family}{See family in glmnet::glmnet function.}

\item{n_iter}{Number of iterations}

\item{alpha}{The alpha parameter for elastic net models. See the alpha argument in glmnet::glmnet.}

\item{n_features}{Breaking the CPOP-Step 1 loop if a certain number of features is reached.}

\item{s}{Method to select a lambda estimate, either "lambda.min" (default) or "lambda.1se".}

\item{cpop1_method}{Default value is "normal". Alternatives are "after" and "either".
\itemize{
\item "normal": meaning that the features selected by **both** data1 (consisted of z1 and y1) and
data2 (consisted of z2 and y2) will be used to construct the final feature set in the first step of CPOP.
\item "after": In case that no predictive features were found to be commonly predictive in both data,
features ever found by **both** data will be pooled to construct the final feature set in the first step of CPOP.
\item "either": In case that no predictive features were found to be commonly predictive in both data,
features ever selected by **either** data will now be pooled
}}

\item{...}{Extra parameter settings for glmnet::cv.glmnet}
}
\value{
A list. Consisted of a vector of features and a tibble of features selected in each step.

A vector of features
}
\description{
Step 1 of the CPOP method, aiming to select features agreed by both data.

Step 1 of the CPOP method, for multiple alpha inputs
}
