% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop1.R
\name{cpop1_iterate}
\alias{cpop1_iterate}
\title{Step 1 of the CPOP method, iteratred}
\usage{
cpop1_iterate(z1, z2, y1, y2, w, nIter = 20, alpha = 1,
  n_features = 50, family = "binomial", s = "lambda.min", ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{w}{A vector}

\item{nIter}{Number of iterations}

\item{alpha}{Lasso alpha, vector}

\item{n_features}{n_features desired}

\item{family}{see glmnet family}

\item{s}{CV-Lasso lambda}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 1 of the CPOP method, for multiple alpha inputs
}
\examples{
data(cpop_data_binary, package = 'CPOP')
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
alpha = c(1, 0.1, 0.01)
cpop1_result = cpop1_iterate(z1, z2, y1, y2, w, nIter = 20,
alpha = alpha, n_features = 30, s = "lambda.min")
cpop1_result
}
