% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop2_escv.R
\name{cpop2_hdci}
\alias{cpop2_hdci}
\title{Step 2 of the CPOP method using HDCI package}
\usage{
cpop2_hdci(z1, z2, y1, y2, cpop1_result, s = "lambda.min", nIter = 20,
  family = "binomial", cpop2_break = TRUE, ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{cpop1_result}{cpop1 result}

\item{s}{CV-Lasso lambda}

\item{nIter}{Number of iterations}

\item{family}{see glmnet family
differential betas are removed}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 2 of the CPOP method using HDCI package
}
\examples{
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
alpha = 0.1
cpop1_result = cpop1_iterate(z1, z2, y1, y2, w, nIter = 20,
alpha = alpha, n_features = 30, s = "lambda.min")
cpop2_hdci_result = cpop2_hdci(z1, z2, y1, y2,
cpop1_result = cpop1_result,
family = "binomial", s = "lambda.min", nIter = 20, cpop2_break = FALSE)
}
