% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop2.R
\name{cpop2_mag}
\alias{cpop2_mag}
\title{Step 2 of the CPOP method}
\usage{
cpop2_mag(z1, z2, y1, y2, family, cpop1_result, s = "lambda.min",
  nIter = 20, cpop2_break = FALSE, mag = 1, ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{family}{see glmnet family}

\item{cpop1_result}{cpop1 result}

\item{s}{CV-Lasso lambda}

\item{nIter}{Number of iterations}

\item{cpop2_break}{If break is needed, logical}

\item{mag}{a threshold
differential betas are removed}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 2 of the CPOP method
}
\examples{
data(cpop_data_binary, package = 'CPOP')
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
alpha = c(1, 0.1, 0.01)
cpop1_result = cpop1_iterate(z1, z2, y1, y2, w, nIter = 20,
alpha = alpha, n_features = 30, s = "lambda.min", family = "binomial")
cpop2_result = cpop2_mag(z1, z2, y1, y2,
cpop1_result = cpop1_result, s = "lambda.min", nIter = 20, family = "binomial")
}
