% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop3.R
\name{cpop3}
\alias{cpop3}
\title{Step 3 of the CPOP method}
\usage{
cpop3(z1, z2, y1, y2, cpop2_result, family, intercept, ...)
}
\arguments{
\item{z1}{A data matrix, columns are pairwise-differences between the original data columns.}

\item{z2}{A data matrix, columns are pairwise-differences between the original data columns.
Column names should be identical to z1.}

\item{y1}{A vector of response variable. Same length as the number of rows of z1.}

\item{y2}{A vector of response variable. Same length as the number of rows of z2.}

\item{cpop2_result}{cpop2 result}

\item{family}{see glmnet family}

\item{intercept}{default to FALSE}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 3 of the CPOP method
}
