% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop3.R
\name{cpop3_hdci}
\alias{cpop3_hdci}
\title{Step 3 of the CPOP method, HDCI}
\usage{
cpop3_hdci(z1, z2, y1, y2, cpop2_result, intercept = FALSE,
  family = "guassian", ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{cpop2_result}{cpop2 result}

\item{intercept}{default to FALSE}

\item{family}{see glmnet family}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 3 of the CPOP method, HDCI
}
\examples{
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
alpha = c(1, 0.1, 0.01)
cpop1_result = cpop1_iterate(z1, z2, y1, y2, w, nIter = 20,
alpha = alpha, n_features = 30, s = "lambda.min", family = "binomial")
cpop2_hdci_result = cpop2_hdci(z1, z2, y1, y2,
cpop1_result = cpop1_result, s = "lambda.min",
nIter = 20, family = "binomial", cpop2_break = TRUE)
cpop3_hdci_result = cpop3_hdci(z1, z2, y1, y2,
cpop2_result = cpop2_hdci_result, family = "binomial", intercept = FALSE)
plot(glmnet::predict.cv.glmnet(cpop3_hdci_result$en1, newx = z1[,cpop2_hdci_result], s = cpop3_hdci_result$en1$lambda.escv),
glmnet::predict.cv.glmnet(cpop3_hdci_result$en2, newx = z1[,cpop2_hdci_result], s = cpop3_hdci_result$en1$lambda.escv))
abline(a = 0, b = 1, col = "red")

}
