% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratio_network.R
\name{plot_lratio_network}
\alias{plot_lratio_network}
\alias{lratio_to_network}
\alias{mst_lratio}
\title{Plot a ratio network}
\usage{
plot_lratio_network(x, type = "ggraph")

lratio_to_network(x)

mst_lratio(x)
}
\arguments{
\item{x}{A character vector representing edges by separating nodes with "--"
i.e. in the form of "e1--e2"}

\item{type}{Type of graphical visualisation, currently only support one of
"visNetwork" (interactive graph).
\itemize{
\item "ggraph" (default): a non-interactive graph visualisation
using the `ggraph` package.
\item "igraph": a `igraph` object from the `igraph` package.
\item "visNetwork": an interactive graph visualisation using the `visNetwork` package
}}
}
\description{
Plot ratio network
}
\examples{
x = c("X1--X10","X1--X3", "X1--X4","X1--X8","X1--X9","X2--X5","X2--X7","X2--X6", "X8--X10")
plot_lratio_network(x, type = "ggraph")
plot_lratio_network(x, type = "igraph")
plot_lratio_network(x, type = "visNetwork")
}
