% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.cpop.R
\name{CROPS.CPOP}
\alias{CROPS.CPOP}
\title{Runs the CPOP algorithm for a range of penalties. The CPOP algorithm finds the best set of changepoints for fitting a change-in-slope model for a single penalty value.
This function uses the CROPS algorithm with CPOP to find all best sets of segmentations for penalty values in a user specified interval.}
\usage{
CROPS.CPOP(y, sigsquared, min_pen = 5, max_pen = 20, PRINT = TRUE)
}
\arguments{
\item{y}{A vector containing the data.}

\item{sigsquared}{Estimate of variance of the residuals (as used by CPOP).}

\item{min_pen}{Minimum penalty value.}

\item{max_pen}{Maximum penalty value. CPOP will find the best segmentation for all penalty values in the interval [min_pen,max_pen].}

\item{PRINT}{If TRUE, the CROPS algorithm will print its progress.}
}
\value{
The output is a list of segmentations found for different penalty values.
\item{summary}{Table summarising the beta values, the corresponding number of changes and the residual sum of squares.}
\item{segmentations}{A list of the changepoints for each of the different segmentations.}
}
\description{
Runs the CPOP algorithm for a range of penalties. The CPOP algorithm finds the best set of changepoints for fitting a change-in-slope model for a single penalty value.
This function uses the CROPS algorithm with CPOP to find all best sets of segmentations for penalty values in a user specified interval.
}
\examples{
## Simulate data
slope=c(rep(0,49),rep(0.1,50),rep(-0.2,50),rep(0,50)) ## slope
mu=cumsum(c(0,slope)) ## underlying piecewise-linear mean
n=length(mu)
y=rnorm(n,mu) ## data
## run CPOP for penalty values in the interval [log(n),2*log(n)]
out=CROPS.CPOP(y,1,min_pen=log(n),max_pen=2*log(n),PRINT=TRUE) 
out[[1]] ## summary
out[[2]][[1]] ## one optimal set of changepoints

}
\references{
Haynes, Eckley and Fearnhead (2017). Computationally efficient changepoint detection for a range of penalties. Journal of Computational and Graphical Statistics, 26(1), 134-143.

Maidstone, Fearnhead and Letchford (2017), Detecting changes in slope with an L_0 penalty. arXiv:1701.01672
}
