% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\name{estimate}
\alias{estimate}
\alias{estimate,cpop.class-method}
\title{A function for estimating the fit of a cpop model}
\usage{
estimate(object, x, ...)
}
\arguments{
\item{object}{An instance of an S4 class produced by \code{\link{cpop}}.}

\item{x}{Locations at which the fit is to be estimated. Default value is the x locations at which the cpop object was defined.}

\item{...}{Additional arguments.}
}
\description{
Estimates the fit of a cpop model at the specified locations.
}
