% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.cpop.R
\name{estimate.CPOP}
\alias{estimate.CPOP}
\title{Estimates the underlying mean function when fitting change-in-slope model.Given a set of changepoints and a data set, this function finds the best continuous
piecewise-linear function that fits the data in terms of minimising the residual sum of squares. The piecewise-linear function is allowed to change slope only at the specified changepoints.}
\usage{
estimate.CPOP(y, tau)
}
\arguments{
\item{y}{A vector containing the data.}

\item{tau}{A vector of changepoints. The changepoints must be ordered; and can, but do not have to, include 0 as first and n as last changepoint.}
}
\value{
A list with two components:
\item{phi}{A vector of the value of the fitted mean at points the changepoints (the first entry is the value at 0, and the last is the value at n).
           The mean function can then be calculated using linear interpolation.}
\item{f}{A vector of length n, giving the value of the mean at each time-point: 1, 2, ..., n.}
}
\description{
Estimates the underlying mean function when fitting change-in-slope model.Given a set of changepoints and a data set, this function finds the best continuous
piecewise-linear function that fits the data in terms of minimising the residual sum of squares. The piecewise-linear function is allowed to change slope only at the specified changepoints.
}
\examples{
### Simulate data
slope=c(rep(0,49),rep(0.1,50),rep(-0.2,50),rep(0,50)) ## slope
mu=cumsum(c(0,slope)) ## underlying piecewise-linear mean
n=length(mu)
y=rnorm(n,mu) ## data
## run CPOP to estimate changepoints, using 2logn penalty and known sigsquared
out=CPOP(y,beta=2*log(length(y)),sigsquared=1)
## obtain estimate mean for these changepoints
mean.estimate=estimate.CPOP(y,out$changepoints)
## plot the data, and fitted mean
plot(1:n,y,xlab="Time",ylab="",type="l")
lines(1:n,mean.estimate$f,col=2,lwd=2)

}
\references{
Maidstone, Fearnhead and Letchford (2017), Detecting changes in slope with an L_0 penalty. arXiv:1701.01672
}
