% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weights.R
\name{feature_weights}
\alias{feature_weights}
\title{Compute feature-wise concordance weights}
\usage{
feature_weights(x1, x2)
}
\arguments{
\item{x1}{A data matrix}

\item{x2}{A data matrix}
}
\value{
A vector
}
\description{
Compute feature-wise concordance weights
}
\examples{
n = 100
p = 10
x1 = matrix(rnorm(100 * 10, mean = 0, sd = 1), nrow = n, ncol = p)
x2 = matrix(rnorm(100 * 10, mean = 1, sd = 1), nrow = n, ncol = p)
feature_weights(x1, x2)
}
