% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\docType{methods}
\name{fitted}
\alias{fitted}
\alias{fitted,cpop.class-method}
\title{Extract Model Fitted Values}
\arguments{
\item{object}{An instance of an S4 class produced by \code{\link{cpop}}.}
}
\value{
A data frame containing the endpoint coordinates for each line segment fitted between
the detected changepoints. The data frame also contains the gradient and intercept values
for each segment and the corresponding residual sum of squares (RSS)..
}
\description{
Extracts the fitted valus produced by \code{\link{cpop}}.
}
