% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lasso_coef.R
\name{get_lasso_coef}
\alias{get_lasso_coef}
\title{Get cv.glmnet coefficient as a matrix or a tibble}
\usage{
get_lasso_coef(lassoObj, s, tibble = FALSE)
}
\arguments{
\item{lassoObj}{cv.glmnet object}

\item{s}{CV-Lasso lambda}

\item{tibble}{Default to FALSE}
}
\value{
A matrix or a tibble
}
\description{
Get cv.glmnet coefficient as a matrix or a tibble
}
\examples{
set.seed(1)
n = 1000
p = 10
x = matrix(rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
colnames(x) = paste0("X", 1:p)
k = 2
beta = c(rep(1, k), rep(0, p - k))
expit = function(x) 1/(1+exp(-x))
y = rbinom(n, 1, prob = expit(x \%*\% beta))
lassoObj = glmnet::cv.glmnet(x = x, y = y, family = "binomial")
get_lasso_coef(lassoObj = lassoObj, s = "lambda.min")
get_lasso_coef(lassoObj = lassoObj, s = "lambda.min", tibble = TRUE)
}
