% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_escv.R
\name{glmnet_escv}
\alias{glmnet_escv}
\title{Step 1 of the CPOP method}
\usage{
glmnet_escv(x, y, n_folds = 5, lambda = 10^seq(3, -3, length = 100),
  ...)
}
\arguments{
\item{x}{A data matrix}

\item{y}{A vector}

\item{n_folds}{Number of CV folds}

\item{lambda}{Default to 10^seq(2, -3, length = 100)}

\item{...}{extra parameters pass onto glmnet}
}
\value{
A vector
}
\description{
Step 1 of the CPOP method, for a single given alpha
}
\examples{
set.seed(1)
n = 1000
p = 10
x1 = matrix(rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
x2 = x1 + matrix(rnorm(n * p, mean = 0, sd = 0.1), nrow = n, ncol = p)
colnames(x1) = colnames(x2) = sprintf("X\%02d", 1:p)
k = 2
beta = c(rep(1, k), rep(0, p - k))
y1 = x1 \%*\% beta + rnorm(n, mean = 0, sd = 0.5)
y2 = x2 \%*\% beta + rnorm(n, mean = 0, sd = 0.5)
glmnet_escv(x = x1, y = y1)
}
