% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{identity_dist}
\alias{identity_dist}
\alias{identity_dist_matrix}
\alias{calcIdenDist}
\alias{identityDist}
\title{Identity distance between two vectors}
\usage{
identity_dist(x, y, iqr = FALSE)

identity_dist_matrix(matrix)

calcIdenDist(matrix)

identityDist(x, y, iqr = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{iqr}{logical, default to FALSE. If set to TRUE, then IQR of the
identity distance is returned.}

\item{matrix}{a matrix}
}
\description{
Identity distance (`identity_dist`) between two vectors is
defined as: d = (x - y)/sqrt(2). The `identity_dist_matrix` function performs this
calculation for every pairwise rows in a matrix.
}
\examples{
set.seed(1)
x = rnorm(100)
identity_dist(x = x, y = 3*x, iqr = FALSE)
identity_dist(x = x, y = 3*x, iqr = TRUE)
set.seed(1)
x = matrix(rnorm(100), ncol = 20)
identity_dist_matrix(x)
}
