% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{identityDist}
\alias{identityDist}
\title{Identity distance between two vectors}
\usage{
identityDist(x, y, iqr = FALSE)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{iqr}{logical}
}
\description{
Identity distance between two vectors
}
\examples{
set.seed(1)
x = rnorm(100)
identityDist(x = x, y = 3*x, iqr = FALSE)
identityDist(x = x, y = 3*x, iqr = TRUE)
}
