% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_glmnet.R
\name{naive_glmnet}
\alias{naive_glmnet}
\title{Naive glmnet fitting procedure}
\usage{
naive_glmnet(z1, z2, y1, y2, s = "lambda.min", ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{s}{Default to "lambda.min"}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Naive glmnet fitting procedure
}
\examples{
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
s = "lambda.min"
naive_glmnet_result = naive_glmnet(z1, z2, y1, y2,
family = "binomial", alpha = 1, s = "lambda.min")
plot_glmnet_coef(naive_glmnet_result, s = "lambda.min", type = "point")
plot_glmnet_coef(naive_glmnet_result, s = "lambda.min", type = "text")
plot_glmnet_coef(naive_glmnet_result, s = "lambda.min", type = "bar")
}
