% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_panel.R
\name{panel_idenDist}
\alias{panel_idenDist}
\title{pairs panel with identity distance}
\usage{
panel_idenDist(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\description{
pairs panel with identity distance
}
\examples{
pairs(iris[,-5],
upper.panel = panel_scatter_abline,
lower.panel = panel_idenDist)
}
\author{
Kevin Wang
}
