% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_panel.R
\name{panel_scatter_abline_smallDots}
\alias{panel_scatter_abline_smallDots}
\title{A better pairs plot function}
\usage{
panel_scatter_abline_smallDots(x, y, a = 0, b = 1, col = "red")
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}

\item{a}{a for abline}

\item{b}{b for abline}

\item{col}{col for abline}
}
\description{
panel_scatter_abline with small dots
}
\examples{
pairs(iris[,-5],
lower.panel = panel_cor,
upper.panel = panel_scatter_abline_smallDots)
}
\author{
Kevin Wang
}
