% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs_panel.R
\name{panel_scatter_smallDots_abhvline}
\alias{panel_scatter_smallDots_abhvline}
\title{A better pairs plot function}
\usage{
panel_scatter_smallDots_abhvline(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\description{
panel_scatter_smallDots with h and v lines at zero
}
\examples{
pairs(iris[,-5],
lower.panel = panel_cor,
upper.panel = panel_scatter_smallDots_abhvline)
}
\author{
Kevin Wang
}
