% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,cpop.class,ANY-method}
\title{Visualisation of Changepoint Locations and Data}
\usage{
\S4method{plot}{cpop.class,ANY}(x)
}
\arguments{
\item{x}{An instance of an cpop S4 class produced by \code{\link{cpop}}.}
}
\value{
A ggplot object.
}
\description{
Plot methods for S4 objects returned by \code{\link{cpop}}.
}
