% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpop_coef.R
\name{plot_cpop_coef}
\alias{plot_cpop_coef}
\title{Plot cpop coefficients}
\usage{
plot_cpop_coef(cpop_model_result, s = "lambda.min", type = "point")
}
\arguments{
\item{cpop_model_result}{The output of cpop_model}

\item{s}{laso s}

\item{type}{"scatter" or "bar"}
}
\description{
Plot cpop coefficients
}
\examples{
data(cpop_data, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
cpop_model_result = cpop_model(z1, z2, y1, y2, w = w,
alpha = 1, n_features = 40, s = "lambda.min")
alpha = c(1, 0.1)
cpop_model_result = cpop_model(z1, z2, y1, y2, w = w,
cpop1_iterate = TRUE, alpha = alpha, n_features = 40, s = "lambda.min")
plot_cpop_coef(cpop_model_result, s = "lambda.min", type = "point")
plot_cpop_coef(cpop_model_result, s = "lambda.min", type = "text")
plot_cpop_coef(cpop_model_result, s = "lambda.min", type = "bar")
}
