% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glmnet_coef.R
\name{plot_glmnet_coef}
\alias{plot_glmnet_coef}
\title{Plot cpop coefficients}
\usage{
plot_glmnet_coef(cpop_result, s = "lambda.min", type = "point")
}
\arguments{
\item{cpop_result}{The output of cpop_model}

\item{s}{laso s}

\item{type}{"scatter" or "bar"}
}
\description{
Plot cpop coefficients
}
\examples{
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = colmeans_penalty(z1, z2)
cpop_cpop_result = cpop_model(z1, z2, y1, y2, w = w,
alpha = 1, n_features = 40, s = "lambda.min")
plot_glmnet_coef(cpop_cpop_result, s = "lambda.min", type = "point")
plot_glmnet_coef(cpop_cpop_result, s = "lambda.min", type = "text")
plot_glmnet_coef(cpop_cpop_result, s = "lambda.min", type = "bar")
}
