% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratio_network.R
\name{plot_ratio_network}
\alias{plot_ratio_network}
\title{Plot ratio network}
\usage{
plot_ratio_network(x, nodesData = FALSE, stat_df = NULL)
}
\arguments{
\item{x}{The output of top1 or top2 or get_lasso_coef}

\item{nodesData}{Logical}

\item{stat_df}{stat_df}
}
\description{
Plot ratio network
}
\examples{
library(visNetwork)
x = c("X1--X10","X1--X3","X2--X4","X1--X4","X1--X8","X1--X9",
"X10--X2","X2--X5","X2--X7","X2--X6")
plot_ratio_network(x, nodesData = FALSE) \%>\%
visNetwork::visPhysics(solver = "forceAtlas2Based",
forceAtlas2Based = list(gravitationalConstant = -100))
plot_ratio_network(x, nodesData = TRUE)
}
