% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cpop.R
\name{predict_cpop}
\alias{predict_cpop}
\title{Prediction function for CPOP}
\usage{
predict_cpop(cpop_result, newz, s = "lambda.min")
}
\arguments{
\item{cpop_result}{cpop_model result}

\item{newz}{New data}

\item{s}{CV-Lasso lambda}
}
\value{
A vector
}
\description{
Prediction function for CPOP
}
\examples{
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
cpop_result = cpop_model(z1, z2, y1, y2, alpha = 0.1, n_features = 10)
cpop_result
newz = z1[sample(1:nrow(z1), 10),]
head(predict_cpop(cpop_result, newz = newz))
}
