% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_glmnet.R
\name{predict_naive_glmnet}
\alias{predict_naive_glmnet}
\title{Prediction method for naive glmnet}
\usage{
predict_naive_glmnet(glmnet_result, newz, s = "lambda.min")
}
\arguments{
\item{glmnet_result}{glmnet_result}

\item{newz}{matrix}

\item{s}{Default to "lambda.min"}
}
\description{
Prediction method for naive glmnet
}
