% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cpop.R
\name{prep_cpop}
\alias{prep_cpop}
\title{Preparation for CPOP modelling}
\usage{
prep_cpop(x1, x2, force = FALSE)
}
\arguments{
\item{x1}{A data matrix of size n times p}

\item{x2}{A data matrix of size n times p}

\item{force}{If the data matrices has too many columns (p > 200), then the
function will not compute output the
z-matrices (pairwise column differences of x's). Default to FALSE.}
}
\description{
Checks before CPOP modelling, and prepare the z-matrices for CPOP modelling.
}
\examples{
data(cpop_data_binary, package = 'CPOP')
## Loading simulated matrices and vectors
x1 = cpop_data_binary$x1
x2 = cpop_data_binary$x2
prep_cpop(x1 = x1, x2 = x2)
}
