% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\name{residuals}
\alias{residuals}
\alias{residuals,cpop.class-method}
\title{Extracts residuals from a cpop model}
\usage{
\S4method{residuals}{cpop.class}(object)
}
\arguments{
\item{object}{An instance of an S4 class produced by \code{\link{cpop}}.}
}
\description{
Extracts residuals from a cpop model
}
