% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,cpop.class-method}
\title{Displays an S4 object produced by cpop.}
\usage{
show(object)
}
\arguments{
\item{object}{An instance of an S4 class produced by \code{\link{cpop}}.}
}
\description{
Displays an S4 object produced by \code{\link{cpop}}.
}
