% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,cpop.class-method}
\title{Summary of cpop Analysis.}
\usage{
\S4method{summary}{cpop.class}(object)
}
\arguments{
\item{object}{An instance of an S4 class produced by \code{\link{cpop}}.}
}
\description{
Summary method for results produced by \code{\link{cpop}}.
}
