% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top1.R
\name{top1_iterate}
\alias{top1_iterate}
\title{Step 1 of the TOP method, iteratred}
\usage{
top1_iterate(z1, z2, y1, y2, w, nIter = 20, alpha = c(1, 0.1),
  n_features = 50, s = "lambda.min", ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{w}{A vector}

\item{nIter}{Number of iterations}

\item{alpha}{Lasso alpha, vector}

\item{n_features}{n_features desired}

\item{s}{CV-Lasso lambda}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 1 of the TOP method, iteratred
}
\examples{
set.seed(1)
n = 1000
p = 10
x1 = matrix(rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
x2 = x1 + matrix(rnorm(n * p, mean = 0, sd = 0.1), nrow = n, ncol = p)
colnames(x1) = colnames(x2) = paste0("X", 1:p)
k = 2
beta = c(rep(1, k), rep(0, p - k))
expit = function(x) 1/(1+exp(-x))
y1 = rbinom(n, 1, prob = expit(x1 \%*\% beta))
y2 = rbinom(n, 1, prob = expit(x2 \%*\% beta))
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
alpha = c(1, 0.1, 0)
top1_iterate(z1, z2, y1, y2, w, nIter = 20, alpha = alpha, s = "lambda.min")
}
