% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top2.R
\name{top2}
\alias{top2}
\title{Step 2 of the TOP method}
\usage{
top2(z1, z2, y1, y2, top1_result, s = "lambda.min", nIter = 20,
  family = "binomial", ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{top1_result}{top1 result}

\item{s}{CV-Lasso lambda}

\item{nIter}{Number of iterations}

\item{family}{see glmnet family}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 2 of the TOP method
}
\examples{
set.seed(1)
n = 1000
p = 10
x1 = matrix(rnorm(n * n, mean = 0, sd = 1), nrow = n, ncol = p)
x2 = matrix(rnorm(n * p, mean = 1, sd = 1), nrow = n, ncol = p)
colnames(x1) = colnames(x2) = paste0("X", 1:p)
k = 2
beta = c(rep(1, k), rep(0, p - k))
expit = function(x) 1/(1+exp(-x))
y1 = rbinom(n, 1, prob = expit(x1 \%*\% beta))
y2 = rbinom(n, 1, prob = expit(x2 \%*\% beta))
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
nIter = 20
top1_result = top1(z1, z2, y1, y2, w, nIter = 20, alpha = 1, s = "lambda.min")
s = "lambda.min"
top2_result = top2(z1, z2, y1, y2, top1_result = top1_result, s = "lambda.min", nIter = 20)
}
