% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_model.R
\name{top_model}
\alias{top_model}
\title{All three steps of TOP modelling}
\usage{
top_model(z1, z2, y1, y2, w, n_features = 50, nIter = 20, alpha = 1,
  family = "binomial", s = "lambda.min", top2_break = TRUE,
  top2_type = "sign", top2_mag = 1, ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{w}{A vector}

\item{n_features}{n_features desired}

\item{nIter}{Number of iterations}

\item{alpha}{Lasso alpha}

\item{family}{family of glmnet}

\item{s}{CV-Lasso lambda}

\item{top2_break}{Should top2 loop be broken the first time}

\item{top2_type}{"sign" or "mag"}

\item{top2_mag}{a threshold
differential betas are removed}

\item{...}{Extra parameter settings for cv.glmnet in top1}
}
\value{
A vector
}
\description{
All three steps of TOP modelling
}
\examples{
set.seed(1)
n = 100
p = 10
x1 = matrix(rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
x2 = x1 + 0.1
colnames(x1) = colnames(x2) = paste0("X", 1:p)
k = 2
beta = c(rep(1, k), rep(0, p - k))
expit = function(x) 1/(1+exp(-x))
y1 = rbinom(n, 1, prob = expit(x1 \%*\% beta))
y2 = rbinom(n, 1, prob = expit(x2 \%*\% beta))
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
top_model_result = top_model(z1, z2, y1, y2, w = w,
alpha = 1, n_features = 40, s = "lambda.min")
alpha = c(1, 0.1)
top_model_result = top_model(z1, z2, y1, y2, w = w,
top1_iterate = TRUE, alpha = alpha, n_features = 40, s = "lambda.min")
top_model_result2 = top_model(z1, z2, y1, y2, w = w,
top1_iterate = TRUE, alpha = alpha, n_features = 40, s = "lambda.min",
top2_type = "mag", top2_mag = 1)

top_model_result = top_model(z1, z2, round(abs(y1)), round(abs(y2)), w = w,
                             top1_iterate = TRUE, alpha = alpha,
                             n_features = 40, s = "lambda.min",
                             family = "poisson")
}
