% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_tbl.R
\name{boxplot_tbl}
\alias{boxplot_tbl}
\title{Make a simple boxplot plotting tbl}
\usage{
boxplot_tbl(x, index = 2)
}
\arguments{
\item{x}{a matrix or data.frame}

\item{index}{summarisation index, "1" for rows or "2" (default) for columns.}
}
\value{
A tibble
}
\description{
Make a simple boxplot plotting tbl
}
\examples{
data(cpop_data_binary, package = 'CPOP')
## Loading example matrix
x1 = cpop_data_binary$x1
set.seed(1)
boxplot_tbl(x = x1, index = 1)
boxplot_tbl(x = x1, index = 2)
}
