% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\name{CPOP}
\alias{CPOP}
\title{CPOP}
\usage{
CPOP(y, beta, sigsquared = 1)
}
\arguments{
\item{y}{A vector containing the data.}

\item{beta}{A positive real value for the penalty incurred for adding a changepoint (prevents over-fitting).}

\item{sigsquared}{Estimate of residual variance.}
}
\value{
Minimum value of the penalised cost and associated changepoints. The
vector of changepoint locations always starts with 0 and ends with n. A
change at time t corresponds to the slope after time t being different
from the slope prior to time t.
}
\description{
Pruned dynamic programming algorithm for finding the best segmentation
of data for a change-in-slope model where best is defined in terms of
minimising the residual sum of squares plus a penalty that is linear in
the number of changepoints. The penalty for adding each change-point is
beta. The choice of beta is important to the reliability of the
estimates, with larger values of beta leading to fewer estimated
changepoints. If the residual noise is IID Gaussian then a penalty of
\eqn{2log(n)} is a reasonable default. For strongly correlated noise a
larger penalty would be needed.
}
\details{
The residual sum of squares is scaled by the estimate of the residual
variance, sigsquared.
}
\examples{


### Simulate data

slope=c(rep(0,49),rep(0.1,50),rep(-0.2,50),rep(0,50)) ## slope
mu=cumsum(c(0,slope)) ## underlying piecewise-linear mean
n=length(mu)
y=rnorm(n,mu) ## data

## plot data and mean
plot(1:n,y,xlab="Time",ylab="",type="l")
lines(1:n,mu,col=2,lwd=2)

## run CPOP to estimate changepoints, using 2logn penalty and known sigsquared
out=CPOP(y,beta=2*log(length(y)),sigsquared=1)

out$changepoints ## list of estimated changepoints
## plot changepoints -- note that 0,n are included automatically in this list.
abline(v=out$changepoints,col=4) 

## re-run CPOP with too small a penalty
out2=CPOP(y,beta=4,sigsquared=1)
out2$changepoints ## list of estimated changepoints
abline(v=out2$changepoints,col=5) ## plot changepoints -- too many estimated

 
}
\references{
Maidstone, Fearnhead and Letchford (2017), Detecting changes in slope
 with an L_0 penalty. arXiv:1701.01672
}
