% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\name{cpop}
\alias{cpop}
\title{cpop}
\usage{
cpop(
  y,
  x = 1:length(y),
  grid = x,
  beta = 2 * log(length(y)),
  sd = 1,
  minseglen = 0,
  prune.approx = FALSE
)
}
\arguments{
\item{y}{A vector of length n containing the data.}

\item{x}{A vector of length n containing the locations of y. Default value is NULL, in which case the locations \code{x = 1:length(y)} are assumed.}

\item{grid}{An ordered vector of possible locations for the change points. If this is NULL, then this is set to x, the vector of times/locations of the data points.}

\item{beta}{A positive real value for the penalty incurred for adding a changepoint (prevents over-fitting).}

\item{sd}{Estimate of residual standard deviation. Can be a single numerical value or a vector of values for the case of varying standard deviation. Default value is 1.}

\item{minseglen}{The minimum allowable segment length, i.e. distance between successive changepoints. Default is 0.}

\item{prune.approx}{Only relevant if a minimum segment length is set. If True, cpop will use an approximate pruning algorithm that will speed up computation but may
occasionally lead to a sub-optimal solution in terms of the estimate change point locations. If the minimum segment length is 0, then an exact pruning algorithm is possible and is used.}
}
\value{
An instance of an S4 class of type cpop.class.
}
\description{
Algorithm for finding the best segmentation of data for a change-in-slope model.
}
\examples{

library(cpop)
# generate some test data
set.seed(0)
x <- seq(0,1,0.01)
n <- length(x)
sigma <- rep(0.1,n)
mu <- c(2*x[1:floor(n/2)],2 - 2*x[(floor(n/2)+1):n])
y <- rnorm(n,mu,sigma)

# use the locations in x
res <- cpop(y,x,beta=2*log(length(y)),sd=sigma)
plot(res)

# without locations (note explicit paramater names)
res <- cpop(y,beta=2*log(length(y)),sd=sigma)
plot(res)

# stretch the end of the data
x[75:101] <- x[75:101] + seq(from=0,by=0.2,length.out=27)
res <- cpop(y,x,beta=2*log(length(y)),sd=sigma)
plot(res)
 
}
\references{
\insertRef{doi:10.1080/10618600.2018.1512868}{cpop}
}
