% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop.R
\name{cpop}
\alias{cpop}
\title{cpop}
\usage{
cpop(y, x = NULL, beta = 0, sigsquared = 1)
}
\arguments{
\item{y}{A vector of length n containing the data.}

\item{x}{A vector of length n containing the locations of y. Default value is NULL, in which case the locations \code{x = 1:length(y)} are assumed.}

\item{beta}{A positive real value for the penalty incurred for adding a changepoint (prevents over-fitting).}

\item{sigsquared}{Estimate of residual variance. Default value is 1.}
}
\value{
An instance of an S4 class of type cpop.class.
}
\description{
Algorithm for finding the best segmentation of data for a change-in-slope model.
}
\examples{

library(cpop)
# generate some test data
set.seed(0)
x <- seq(0,1,0.01)
n <- length(x)
sigma <- rep(0.1,n)
mu <- c(2*x[1:floor(n/2)],2 - 2*x[(floor(n/2)+1):n])
y <- rnorm(n,mu,sigma)

# use the locations in x
res <- cpop(y,x,2*log(length(y)),0.1)
plot(res)

# without locations (note explicit paramater names)
res <- cpop(y,beta=2*log(length(y)),sigsquared=0.1)
plot(res)

# stretch the end of the data
x[75:101] <- x[75:101] + seq(from=0,by=0.2,length.out=27)
res <- cpop(y,x,2*log(length(y)),0.1)
plot(res)
 
}
\references{
\insertRef{doi:10.1080/10618600.2018.1512868}{cpop}
}
