% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop1.R
\name{cpop1}
\alias{cpop1}
\alias{cpop1_iterate}
\title{Step 1 of the CPOP method}
\usage{
cpop1(
  z1,
  z2,
  y1,
  y2,
  w,
  family,
  n_iter = 20,
  alpha = 1,
  n_features = 50,
  s = "lambda.min",
  cpop1_method = "normal",
  ...
)

cpop1_iterate(
  z1,
  z2,
  y1,
  y2,
  w = NULL,
  family,
  s = "lambda.min",
  n_iter = 20,
  alpha = 1,
  n_features = 50,
  ...
)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{w}{A vector}

\item{family}{See glmnet family}

\item{n_iter}{Number of iterations}

\item{alpha}{Lasso alpha}

\item{n_features}{n_features desired}

\item{s}{CV lambda}

\item{cpop1_method}{Default value is "normal". Alternatives are "after" and "either".}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector of features

A vector of features
}
\description{
Step 1 of the CPOP method, for a single given alpha.
Step 1 of CPOP takes in two training data (x1, y1) and (x2, y2) and performed weighted
elastic net variable selection.

Step 1 of the CPOP method, for multiple alpha inputs
}
