% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop1.R
\name{cpop1}
\alias{cpop1}
\alias{cpop1_iterate}
\title{Step 1 of the CPOP method}
\usage{
cpop1(
  z1,
  z2,
  y1,
  y2,
  w,
  family,
  n_iter = 20,
  alpha = 1,
  n_features = 50,
  s = "lambda.min",
  ...
)

cpop1_iterate(
  z1,
  z2,
  y1,
  y2,
  w,
  family,
  s = "lambda.min",
  n_iter = 20,
  alpha = 1,
  n_features = 50,
  ...
)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{w}{A vector}

\item{family}{See glmnet family}

\item{n_iter}{Number of iterations}

\item{alpha}{Lasso alpha}

\item{n_features}{n_features desired}

\item{s}{CV lambda}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector of features

A vector of features
}
\description{
Step 1 of the CPOP method, for a single given alpha

Step 1 of the CPOP method, for multiple alpha inputs
}
\examples{
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
cpop1(z1 = z1, z2 = z2, y1 = y1, y2 = y2, w = w, family = "binomial", alpha = 1)
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
cpop1_iterate(z1 = z1, z2 = z2, y1 = y1, y2 = y2, w = w, family = "binomial", alpha = c(0.5, 1))
}
