% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop1.R
\name{cpop1}
\alias{cpop1}
\title{Step 1 of the CPOP method}
\usage{
cpop1(z1, z2, y1, y2, w, nIter = 20, alpha = 1, family = "gaussian",
  n_features = 50, s = "lambda.min", ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{w}{A vector}

\item{nIter}{Number of iterations}

\item{alpha}{Lasso alpha}

\item{family}{See glmnet family}

\item{n_features}{n_features desired}

\item{s}{CV-Lasso lambda}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 1 of the CPOP method, for a single given alpha
}
\examples{
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
nIter = 20
alpha = 0.1
s = "lambda.min"
cpop1(z1, z2, y1, y2, w, nIter = 20,
family = "binomial",
n_features = 20, alpha = alpha, s = "lambda.min")
}
