% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop3.R
\name{cpop3}
\alias{cpop3}
\title{Step 3 of the CPOP method}
\usage{
cpop3(z1, z2, y1, y2, cpop2_result, family, intercept = FALSE, ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{cpop2_result}{cpop2 result}

\item{family}{see glmnet family}

\item{intercept}{default to FALSE}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 3 of the CPOP method
}
