% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop3.R
\name{cpop3}
\alias{cpop3}
\title{Step 3 of the CPOP method iterate}
\usage{
cpop3(z1, z2, y1, y2, cpop2_result, intercept = FALSE,
  family = "binomial", ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{cpop2_result}{cpop2 result}

\item{intercept}{default to FALSE}

\item{family}{see glmnet family}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 3 of the CPOP method
}
\examples{
data(cpop_data, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
alpha = c(1, 0.1, 0.01)
cpop1_result = cpop1_iterate(z1, z2, y1, y2, w, nIter = 20,
alpha = alpha, n_features = 30, s = "lambda.min")
cpop2_result = cpop2_sign(z1, z2, y1, y2, cpop1_result = cpop1_result, s = "lambda.min", nIter = 20)
cpop3_result = cpop3(z1, z2, y1, y2, cpop2_result = cpop2_result, intercept = FALSE)
plot(predict(cpop3_result$en1, newx = z1[,cpop2_result], s = "lambda.min"),
predict(cpop3_result$en2, newx = z1[,cpop2_result], s = "lambda.min"))
abline(a = 0, b = 1)

}
