% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop3_escv.R
\name{cpop3_hdci}
\alias{cpop3_hdci}
\title{Step 3 of the CPOP method, HDCI}
\usage{
cpop3_hdci(z1, z2, y1, y2, cpop2_result, family, intercept = FALSE, ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{cpop2_result}{cpop2 result}

\item{family}{see glmnet family}

\item{intercept}{default to FALSE}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Step 3 of the CPOP method, HDCI
}
\examples{
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
cpop1_result = cpop1_iterate(
z1 = z1, z2 = z2, y1 = y1, y2 = y2, w = w,
family = "binomial", alpha = 0.1)
cpop2_hdci_result = cpop2_hdci(z1 = z1, z2 = z2, y1 = y1, y2 = y2,
cpop1_result = cpop1_result, family = "binomial")
cpop3_hdci_result = cpop3_hdci(z1, z2, y1, y2,
cpop2_result = cpop2_hdci_result, family = "binomial", intercept = FALSE)
}
