% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop_model.R
\name{cpop_model}
\alias{cpop_model}
\title{All three steps of CPOP modelling}
\usage{
cpop_model(
  z1,
  z2,
  y1,
  y2,
  w = NULL,
  n_features = 50,
  n_iter = 20,
  alpha = 1,
  family = "binomial",
  s = "lambda.min",
  cpop2_break = TRUE,
  cpop2_type = "sign",
  cpop2_mag = 1,
  cpop1_method = "normal",
  intercept = FALSE,
  ...
)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{w}{A vector}

\item{n_features}{n_features desired}

\item{n_iter}{Number of iterations}

\item{alpha}{Lasso alpha}

\item{family}{family of glmnet}

\item{s}{CV-Lasso lambda}

\item{cpop2_break}{Should cpop2 loop be broken the first time}

\item{cpop2_type}{"sign" or "mag"}

\item{cpop2_mag}{a threshold}

\item{cpop1_method}{Default value is "normal". Alternatives are "after" and "either".
differential betas are removed}

\item{intercept}{Default to FALSE}

\item{...}{Extra parameter settings for cv.glmnet in cpop1}
}
\value{
A vector
}
\description{
All three steps of CPOP modelling
}
\examples{
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
cpop_result = cpop_model(z1, z2, y1, y2, alpha = 1, n_features = 10)
cpop_result
cpop_result = cpop_model(z1, z2, y1, y2, alpha = 1, n_features = 10, cpop1_method = "after")
}
