% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop_model.R
\name{cpop_model}
\alias{cpop_model}
\title{All three steps of CPOP modelling}
\usage{
cpop_model(z1, z2, y1, y2, w, n_features = 50, nIter = 20, alpha = 1,
  family = "binomial", s = "lambda.min", cpop2_break = TRUE,
  cpop2_type = "sign", cpop2_mag = 1, ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{w}{A vector}

\item{n_features}{n_features desired}

\item{nIter}{Number of iterations}

\item{alpha}{Lasso alpha}

\item{family}{family of glmnet}

\item{s}{CV-Lasso lambda}

\item{cpop2_break}{Should cpop2 loop be broken the first time}

\item{cpop2_type}{"sign" or "mag"}

\item{cpop2_mag}{a threshold
differential betas are removed}

\item{...}{Extra parameter settings for cv.glmnet in cpop1}
}
\value{
A vector
}
\description{
All three steps of CPOP modelling
}
\examples{
data(cpop_data, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
cpop_model_result = cpop_model(z1, z2, y1, y2, w = w,
alpha = 1, n_features = 40, s = "lambda.min")
alpha = c(1, 0.1)
cpop_model_result = cpop_model(z1, z2, y1, y2, w = w,
cpop1_iterate = TRUE, alpha = alpha, n_features = 40, s = "lambda.min")
}
