% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpop_model.R
\name{cpop_model}
\alias{cpop_model}
\title{CPOP modelling}
\usage{
cpop_model(
  z1,
  z2,
  y1,
  y2,
  w = NULL,
  n_features = 50,
  n_iter = 20,
  alpha = 1,
  family = "binomial",
  s = "lambda.min",
  cpop2_break = TRUE,
  cpop2_type = "sign",
  cpop2_mag = 1,
  cpop1_method = "normal",
  intercept = FALSE,
  ...
)
}
\arguments{
\item{z1}{A data matrix, columns are pairwise-differences between the original data columns.}

\item{z2}{A data matrix, columns are pairwise-differences between the original data columns.
Column names should be identical to z1.}

\item{y1}{A vector of response variable. Same length as the number of rows of z1.}

\item{y2}{A vector of response variable. Same length as the number of rows of z2.}

\item{w}{A vector of weights.}

\item{n_features}{Breaking the CPOP-Step 1 loop if a certain number of features is reached.}

\item{n_iter}{Number of iterations in Step 1 and 2.}

\item{alpha}{The alpha parameter for elastic net models. See the alpha argument in glmnet::glmnet.}

\item{family}{family of glmnet}

\item{s}{CV-Lasso lambda}

\item{cpop2_break}{Should cpop2 loop be broken the first time}

\item{cpop2_type}{"sign" or "mag"}

\item{cpop2_mag}{a threshold
differential betas are removed}

\item{intercept}{Default to FALSE}

\item{...}{Extra parameter settings for cv.glmnet in cpop1}
}
\value{
A vector
}
\description{
CPOP is consisted of three steps. Step 1 is to select features common to
two transformed data. Note the input must be pairwise-differences between the original data columns.
Step 2 is to select features in constructed models that shared similar characteristics. Step 3 is to
construct a final model used for prediction.
}
\examples{
data(cpop_data_binary, package = 'CPOP')
attach(cpop_data_binary)
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
cpop_result = cpop_model(z1, z2, y1, y2, alpha = 1, n_features = 10)
cpop_result
}
