% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_escv.R
\name{glmnet_escv}
\alias{glmnet_escv}
\title{Step 1 of the CPOP method}
\usage{
glmnet_escv(x, y, n_folds = 5, lambda = 10^seq(3, -3, length = 100),
  ...)
}
\arguments{
\item{x}{A data matrix}

\item{y}{A vector}

\item{n_folds}{Number of CV folds}

\item{lambda}{Default to 10^seq(2, -3, length = 100)}

\item{...}{extra parameters pass onto glmnet}
}
\value{
A vector
}
\description{
Step 1 of the CPOP method, for a single given alpha
}
\examples{
data(cpop_data_linear, package = 'CPOP')
glmnet_escv(x = x1, y = y1)
}
