% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cpop.R
\name{impute_cpop}
\alias{impute_cpop}
\title{Imputing gene expression values using CPOP model}
\usage{
impute_cpop(cpop_result, x1, x2, newx)
}
\arguments{
\item{cpop_result}{cpop_model result}

\item{x1}{gene data 1}

\item{x2}{gene data 2}

\item{newx}{gene gene data}
}
\value{
A vector
}
\description{
Imputing gene expression values using CPOP model
}
\examples{
data(cpop_data_binary, package = 'CPOP')
attach(cpop_data_binary)
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
cpop_result = cpop_model(z1, z2, y1, y2, alpha = 0.1, n_features = 10)
cpop_result
z3 = pairwise_col_diff(x3)
head(predict_cpop(cpop_result, newz = z3))
x4 = x3
x4[,1] = NA
# z4 = pairwise_col_diff(x4)
## head(predict_cpop(cpop_result, newz = z4))
x4_imp = impute_cpop(cpop_result, x1 = x1, x2 = x2, newx = x4)
# head(predict_cpop(cpop_result, newz = CPOP::pairwise_col_diff(x4_imp)))
## plot(predict_cpop(cpop_result, newz = CPOP::pairwise_col_diff(x4_imp))$cpop_model_avg,
## predict_cpop(cpop_result, newz = z3)$cpop_model_avg)
}
