% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_glmnet.R
\name{naive_glmnet}
\alias{naive_glmnet}
\title{Naive glmnet fitting procedure}
\usage{
naive_glmnet(
  x1,
  x2,
  y1,
  y2,
  s = "lambda.min",
  family = "binomial",
  z1,
  z2,
  ...
)
}
\arguments{
\item{y1}{A vector}

\item{y2}{A vector}

\item{s}{Default to "lambda.min"}

\item{family}{family of glmnet}

\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Naive glmnet fitting procedure
}
\examples{
data(cpop_data_binary, package = 'CPOP')
## Loading simulated matrices and vectors
x1 = cpop_data_binary$x1
x2 = cpop_data_binary$x2
y1 = cpop_data_binary$y1
y2 = cpop_data_binary$y2
set.seed(1)
cpop_result = cpop_model(x1 = x1, x2 = x2, y1 = y1, y2 = y2, alpha = 1, n_features = 10)
lasso_result = naive_glmnet(x1 = x1, x2 = x2, y1 = y1, y2 = y2, alpha = 1, intercept = FALSE)
cpop_result
lasso_result
plot_cpop(cpop_result)
plot_cpop(lasso_result)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
plot(predict_cpop(cpop_result, newz = z1)$cpop_model_avg,
predict_naive_glmnet(lasso_result, newz = z1)$naive_glmnet_avg)
abline(a = 0, b = 1, col = "red")
}
