% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_glmnet.R
\name{naive_glmnet}
\alias{naive_glmnet}
\title{Naive glmnet fitting procedure}
\usage{
naive_glmnet(z1, z2, y1, y2, s = "lambda.min", ...)
}
\arguments{
\item{z1}{A data matrix}

\item{z2}{A data matrix}

\item{y1}{A vector}

\item{y2}{A vector}

\item{s}{Default to "lambda.min"}

\item{...}{Extra parameter settings for cv.glmnet}
}
\value{
A vector
}
\description{
Naive glmnet fitting procedure
}
\examples{
data(cpop_data_binary, package = 'CPOP')
attach(cpop_data_binary)
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
cpop_result = cpop_model(z1, z2, y1, y2, alpha = 1, n_features = 10)
lasso_result = naive_glmnet(z1, z2, y1, y2, alpha = 1, intercept = FALSE)
cpop_result
lasso_result
plot_cpop(cpop_result)
plot_cpop(lasso_result)
plot(predict_cpop(cpop_result, newz = z1)$cpop_model_avg,
predict_naive_glmnet(lasso_result, newz = z1)$naive_glmnet_avg)
abline(a = 0, b = 1, col = "red")
}
