% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_col_diff.R
\name{pairwise_col_diff}
\alias{pairwise_col_diff}
\title{Compare pairwise diff}
\usage{
pairwise_col_diff(x)
}
\arguments{
\item{x}{A data matrix}
}
\value{
A pairwise difference matrix
}
\description{
Compare pairwise diff
}
\examples{
n = 4
p = 4
x = matrix(rep(1:p, n), nrow = n, ncol = p, byrow = TRUE)
colnames(x) = paste0("X", 1:p)
identical(pairwise_col_diff(x),
pairwise_col_diff(x[,4:1]))

}
