% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalties.R
\name{colmeans_penalty}
\alias{colmeans_penalty}
\alias{p_transform}
\title{Compute feature-wise concordance penalty}
\usage{
colmeans_penalty(x1, x2)

p_transform(penalty)
}
\arguments{
\item{x1}{A data matrix}

\item{x2}{A data matrix}

\item{penalty}{A vector of penalties to be transformed}
}
\value{
A vector
}
\description{
Compute feature-wise concordance penalty
}
\examples{
n = 10
p = 10
x1 = matrix(rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
x2 = matrix(rnorm(n * p, mean = 1, sd = 1), nrow = n, ncol = p)
colmeans_penalty(x1, x2)
}
