% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratio_network.R
\name{plot_ratio_network}
\alias{plot_ratio_network}
\title{Plot ratio network}
\usage{
plot_ratio_network(x, nodesData = FALSE, stat_df = NULL)
}
\arguments{
\item{x}{The output of top1 or top2 or get_lasso_coef}

\item{nodesData}{Logical}

\item{stat_df}{stat_df}
}
\description{
Plot ratio network
}
\examples{
x = c("X1--X2","X1--X3","X1--X4","X1--X5","X1--X6","X1--X7","X1--X8",
"X1--X9","X1--X10","X2--X3","X2--X4","X2--X5","X2--X6","X2--X7",
"X2--X8","X2--X9","X2--X10","X3--X4","X3--X7","X3--X8","X3--X9",
"X4--X5","X4--X7","X4--X9","X5--X10","X7--X8","X7--X9","X8--X9")
plot_ratio_network(x, nodesData = FALSE)
plot_ratio_network(x, nodesData = TRUE)
}
