% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cpop.R
\name{predict_cpop}
\alias{predict_cpop}
\title{Prediction function for CPOP}
\usage{
predict_cpop(cpop_result, newz, s = "lambda.min", model_number = 1L,
  tibble = FALSE)
}
\arguments{
\item{cpop_result}{cpop_model result}

\item{newz}{New data}

\item{s}{CV-Lasso lambda}

\item{model_number}{1 or 2 or "both"}

\item{tibble}{Logical}
}
\value{
A vector
}
\description{
Prediction function for CPOP
}
\examples{
data(cpop_data_binary, package = 'CPOP')
set.seed(1)
z1 = pairwise_col_diff(x1)
z2 = pairwise_col_diff(x2)
w = compute_weights(z1, z2)
cpop_result = cpop_model(z1, z2, y1, y2, w = w,
alpha = 1, n_features = 40, s = "lambda.min", family = "binomial")
head(predict_cpop(cpop_result, newz = z1, model_number = 1))
head(predict_cpop(cpop_result, newz = z1, model_number = "both", tibble = TRUE))
head(predict_cpop(cpop_result, newz = z1, model_number = "avg", tibble = TRUE))
}
