% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cpop.R
\name{predict_cpop}
\alias{predict_cpop}
\title{Prediction function for CPOP}
\usage{
predict_cpop(cpop_result, newx, s = "lambda.min", newz)
}
\arguments{
\item{cpop_result}{cpop_model result}

\item{newx}{New data, n times p, as original features,}

\item{s}{CV-Lasso lambda}

\item{newz}{(Deprecated) new data, n times choose(p, 2), as ratio features.}
}
\value{
A vector
}
\description{
Prediction function for CPOP
}
\examples{
data(cpop_data_binary, package = 'CPOP')
## Loading simulated matrices and vectors
x1 = cpop_data_binary$x1
x2 = cpop_data_binary$x2
x3 = cpop_data_binary$x3
y1 = cpop_data_binary$y1
y2 = cpop_data_binary$y2
y3 = cpop_data_binary$y3
set.seed(1)
cpop_result = cpop_model(x1 = x1, x2 = x2, y1 = y1, y2 = y2, alpha = 0.1, n_features = 10)
cpop_result
head(predict_cpop(cpop_result, newx = x3))
}
