% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cpop.R
\name{prep_cpop}
\alias{prep_cpop}
\title{Preparation for CPOP modelling}
\usage{
prep_cpop(x1, x2)
}
\arguments{
\item{x1}{A data matrix}

\item{x2}{A data matrix}
}
\description{
Checks before CPOP modelling
}
\examples{
n = 10
p = 10
x1 = matrix(rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
x2 = matrix(rnorm(n * p, mean = 0, sd = 1), nrow = n, ncol = p)
prep_cpop(x1 = x1, x2 = x2)
}
